+++ db/structure.sql
@@ -51,10 +51,10 @@ CREATE TABLE base_slots (
     created_at timestamp without time zone,
     updated_at timestamp without time zone,
     deleted_at timestamp without time zone,
-    meta_slot_id integer,
-    id integer NOT NULL,
+    meta_slot_id bigint NOT NULL,
+    id bigint NOT NULL,
     share_id character varying(8) DEFAULT ''::character varying,
-    shared_by_id integer,
+    shared_by_id bigint,
     slot_type integer NOT NULL
 );
 
@@ -83,9 +83,9 @@ ALTER SEQUENCE base_slots_id_seq OWNED BY base_slots.id;
 --
 
 CREATE TABLE comments (
-    id integer NOT NULL,
-    user_id integer,
-    slot_id integer,
+    id bigint NOT NULL,
+    user_id bigint NOT NULL,
+    slot_id bigint NOT NULL,
     content text,
     deleted_at timestamp without time zone,
     created_at timestamp without time zone,
@@ -118,9 +118,9 @@ ALTER SEQUENCE comments_id_seq OWNED BY comments.id;
 
 CREATE TABLE connects (
     id integer NOT NULL,
-    user_id integer NOT NULL,
+    user_id bigint NOT NULL,
     provider_id integer NOT NULL,
-    social_id integer NOT NULL,
+    social_id bigint NOT NULL,
     social_data hstore DEFAULT ''::hstore,
     deleted_at timestamp without time zone,
     created_at timestamp without time zone,
@@ -152,9 +152,9 @@ ALTER SEQUENCE connects_id_seq OWNED BY connects.id;
 --
 
 CREATE TABLE friendships (
-    id integer NOT NULL,
-    user_id integer,
-    friend_id integer,
+    id bigint NOT NULL,
+    user_id bigint NOT NULL,
+    friend_id bigint NOT NULL,
     state bit(2) DEFAULT B'00'::"bit",
     deleted_at timestamp without time zone,
     created_at timestamp without time zone,
@@ -186,10 +186,10 @@ ALTER SEQUENCE friendships_id_seq OWNED BY friendships.id;
 --
 
 CREATE TABLE group_slots (
-    group_id integer,
+    group_id bigint NOT NULL,
     created_at timestamp without time zone,
     updated_at timestamp without time zone,
-    meta_slot_id integer,
+    meta_slot_id bigint,
     deleted_at timestamp without time zone
 )
 INHERITS (base_slots);
@@ -200,9 +200,9 @@ INHERITS (base_slots);
 --
 
 CREATE TABLE groups (
-    id integer NOT NULL,
-    owner_id integer,
-    name character varying(255),
+    id bigint NOT NULL,
+    owner_id bigint NOT NULL,
+    name character varying(255) NOT NULL,
     members_can_post boolean DEFAULT true,
     members_can_invite boolean DEFAULT false,
     created_at timestamp without time zone,
@@ -235,9 +235,9 @@ ALTER SEQUENCE groups_id_seq OWNED BY groups.id;
 --
 
 CREATE TABLE likes (
-    id integer NOT NULL,
-    user_id integer,
-    base_slot_id integer,
+    id bigint NOT NULL,
+    user_id bigint NOT NULL,
+    base_slot_id bigint NOT NULL,
     deleted_at timestamp without time zone,
     created_at timestamp without time zone,
     updated_at timestamp without time zone
@@ -268,13 +268,13 @@ ALTER SEQUENCE likes_id_seq OWNED BY likes.id;
 --
 
 CREATE TABLE media_items (
-    id integer NOT NULL,
+    id bigint NOT NULL,
     media_type character varying(255),
     public_id character varying(255),
     "position" integer,
     created_at timestamp without time zone,
     updated_at timestamp without time zone,
-    mediable_id integer,
+    mediable_id bigint,
     mediable_type character varying(255),
     deleted_at timestamp without time zone,
     duration integer,
@@ -306,9 +306,9 @@ ALTER SEQUENCE media_items_id_seq OWNED BY media_items.id;
 --
 
 CREATE TABLE memberships (
-    id integer NOT NULL,
-    user_id integer,
-    group_id integer,
+    id bigint NOT NULL,
+    user_id bigint NOT NULL,
+    group_id bigint NOT NULL,
     notifications boolean DEFAULT true,
     created_at timestamp without time zone,
     updated_at timestamp without time zone,
@@ -342,15 +342,15 @@ ALTER SEQUENCE memberships_id_seq OWNED BY memberships.id;
 --
 
 CREATE TABLE meta_slots (
-    id integer NOT NULL,
-    title character varying(48),
-    start_date timestamp without time zone,
-    end_date timestamp without time zone,
-    creator_id integer,
+    id bigint NOT NULL,
+    title character varying(48) NOT NULL,
+    start_date timestamp without time zone NOT NULL,
+    end_date timestamp without time zone NOT NULL,
+    creator_id bigint NOT NULL,
     deleted_at timestamp without time zone,
     created_at timestamp without time zone,
     updated_at timestamp without time zone,
-    location_id integer
+    location_id bigint
 );
 
 
@@ -378,8 +378,8 @@ ALTER SEQUENCE meta_slots_id_seq OWNED BY meta_slots.id;
 --
 
 CREATE TABLE notes (
-    id integer NOT NULL,
-    base_slot_id integer,
+    id bigint NOT NULL,
+    base_slot_id bigint NOT NULL,
     title character varying(255) NOT NULL,
     content text DEFAULT ''::text,
     deleted_at timestamp without time zone,
@@ -412,13 +412,13 @@ ALTER SEQUENCE notes_id_seq OWNED BY notes.id;
 --
 
 CREATE TABLE re_slots (
-    predecessor_id integer NOT NULL,
+    predecessor_id bigint NOT NULL,
     created_at timestamp without time zone,
     updated_at timestamp without time zone,
     deleted_at timestamp without time zone,
-    meta_slot_id integer NOT NULL,
-    slotter_id integer NOT NULL,
-    parent_id integer NOT NULL
+    meta_slot_id bigint,
+    slotter_id bigint NOT NULL,
+    parent_id bigint NOT NULL
 )
 INHERITS (base_slots);
 
@@ -437,13 +437,13 @@ CREATE TABLE schema_migrations (
 --
 
 CREATE TABLE slot_settings (
-    user_id integer,
-    meta_slot_id integer,
+    user_id bigint NOT NULL,
+    meta_slot_id bigint NOT NULL,
     alerts bit(10) DEFAULT B'0000000000'::"bit",
     deleted_at timestamp without time zone,
     created_at timestamp without time zone,
     updated_at timestamp without time zone,
-    id integer NOT NULL
+    id bigint NOT NULL
 );
 
 
@@ -474,8 +474,8 @@ CREATE TABLE std_slots (
     created_at timestamp without time zone,
     updated_at timestamp without time zone,
     deleted_at timestamp without time zone,
-    meta_slot_id integer,
-    owner_id integer
+    meta_slot_id bigint,
+    owner_id bigint NOT NULL
 )
 INHERITS (base_slots);
 
@@ -485,7 +485,7 @@ INHERITS (base_slots);
 --
 
 CREATE TABLE users (
-    id integer NOT NULL,
+    id bigint NOT NULL,
     username character varying(50),
     created_at timestamp without time zone,
     updated_at timestamp without time zone,
@@ -502,10 +502,10 @@ CREATE TABLE users (
     default_friends_public_alerts bit(10) DEFAULT B'0000000000'::"bit",
     default_reslot_alerts bit(10) DEFAULT B'0000000000'::"bit",
     phone integer,
-    location_id integer,
+    location_id bigint,
     public_url character varying(255),
     push boolean DEFAULT true,
-    slot_default_location_id integer,
+    slot_default_location_id bigint,
     slot_default_duration integer,
     slot_default_type_id integer
 );
@@ -1059,4 +1059,6 @@ INSERT INTO schema_migrations (version) VALUES ('20150428150031');
 INSERT INTO schema_migrations (version) VALUES ('20150429103206');
 
 INSERT INTO schema_migrations (version) VALUES ('20150429104404');
+
+INSERT INTO schema_migrations (version) VALUES ('20150429205903');