+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 #Changelog
 
 ##Changes since version from 27.4.15
+- email is not requiered anymore for user creation
+
+- new endpoints:
+
+- changed type of primary key to bigint for users, friendships, groups,
+  memberships, comments, likes, location, media_items, slot_settings,  metaslots and slots
+
+TODO:
+- add creator for notes and media items
+- rename base_slot_id to slot_id in likes, notes